Use lowdb
