--- electron/main/database/sqlite/Readme.md
@@ -1,3 +1,0 @@
-# Node-Sqlite3 Driver for kysely
-
-See https://github.com/koskimas/kysely/issues/128
